feat: upgrade Poseidon constants and LFSR APIs
